requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
wokwi-client>=0.1.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.0.0
websockets>=13.0
pyserial>=3.5
esptool>=4.0
platformio>=6.0
